Authorization

+++ app/Http/Controllers/TaskListController.php
@@ -2,6 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\TaskStatus;
+use App\Http\Resources\TaskListResource;
+use App\Http\Resources\TaskResource;
+use App\Models\Task;
 use App\Models\TaskList;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -14,7 +18,7 @@ class TaskListController extends Controller
     public function index()
     {
         return $this->success([
-            'task_lists' => TaskList::where('user_id', Auth::id())->get()
+            'task_lists' => TaskListResource::collection(TaskList::where('user_id', Auth::id())->get())
         ]);
     }
 
@@ -33,7 +37,7 @@ class TaskListController extends Controller
         $taskList->save();
 
         return $this->success([
-            'task_list' => $taskList
+            'task_list' => new TaskListResource($taskList)
         ]);
     }
 
@@ -42,8 +46,41 @@ class TaskListController extends Controller
      */
     public function show(TaskList $taskList)
     {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+        
         return $this->success([
-            'task_list' => $taskList
+            'task_list' => new TaskListResource($taskList)
+        ]);
+    }
+
+    public function tasks(TaskList $taskList)
+    {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
+        return $this->success([
+            'tasks' => TaskResource::collection($taskList->tasks)
+        ]);
+    }
+
+    public function addTask(Request $request, TaskList $taskList)
+    {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
+        $request->validate([
+            'title' => ['required', 'string'],
+            'details' => ['nullable', 'string']
+        ]);
+
+        $task = new Task();
+        $task->user_id = Auth::id();
+        $task->task_list_id = $taskList->id;
+        $task->status = TaskStatus::ToDo->value;
+        $task->title = $request->title;
+        $task->details = $request->details;
+        $task->save();
+
+        return $this->success([
+            'tasks' => new TaskResource($task)
         ]);
     }
 
@@ -52,6 +89,8 @@ class TaskListController extends Controller
      */
     public function update(Request $request, TaskList $taskList)
     {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
         $request->validate([
             'title' => ['required', 'string']
         ]);
@@ -60,7 +99,7 @@ class TaskListController extends Controller
         $taskList->save();
 
         return $this->success([
-            'task_list' => $taskList
+            'task_list' => new TaskListResource($taskList)
         ]);
     }
 
@@ -69,6 +108,8 @@ class TaskListController extends Controller
      */
     public function destroy(TaskList $taskList)
     {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
         $taskList->delete();
 
         return $this->success();
